add reusable state shell and bodies
