+++ pages/projects/nigeria/states/[slug]/index.tsx
@@ -1,13 +1,23 @@
 import { useRouter } from "next/router";
+import { useEffect } from "react";
 import Link from "next/link";
-import StateDetailsCarousel from "@/components/StateDetailsCarousel";
-
-const ORDER = ["niger", "kwara", "plateau"];
+import StatePageShell from "@/components/StatePageShell";
+import StateProjectBody from "@/components/StateProjectBody";
+import { getStateBySlug } from "@/data/states";
 
 export default function NigeriaStatePage() {
   const router = useRouter();
   const { slug } = router.query;
-  const startIndex = typeof slug === "string" ? ORDER.indexOf(slug) : 0;
+
+  const state = typeof slug === "string" ? getStateBySlug(slug) : undefined;
+
+  useEffect(() => {
+    if (typeof slug === "string" && !state) {
+      router.replace(`/projects/nigeria/states/${slug}/facts`);
+    }
+  }, [state, slug, router]);
+
+  if (typeof slug !== "string" || !state) return null;
 
   return (
     <>
@@ -19,11 +29,14 @@ export default function NigeriaStatePage() {
           <span className="mr-2">←</span> Back to Projects
         </Link>
       </header>
-      <main className="max-w-6xl mx-auto px-4 sm:px-6 pt-16 md:pt-20">
-        <StateDetailsCarousel
-          states={ORDER}
-          startIndex={startIndex >= 0 ? startIndex : 0}
-        />
+      <main className="max-w-6xl mx-auto px-4 sm:px-6 pt-16 md:pt-20 space-y-6">
+        <StatePageShell
+          slug={slug}
+          status={state.status}
+          showFactsLink={Boolean(state.facts && state.facts.length)}
+        >
+          <StateProjectBody state={state} />
+        </StatePageShell>
       </main>
     </>
   );
